Release 2.20

+++ source/Core/Drivers/Utils.h
@@ -10,7 +10,7 @@
 #include <stdint.h>
 class Utils {
 public:
-  static int32_t InterpolateLookupTable(const uint16_t *lookupTable, const int noItems, const uint16_t value);
+  static int32_t InterpolateLookupTable(const int32_t *lookupTable, const int noItems, const int32_t value);
   static int32_t LinearInterpolate(int32_t x1, int32_t y1, int32_t x2, int32_t y2, int32_t x);
 };
